pyproject.toml
[project]
name = "media-tool"
version = "0.1.0"
description = "自动化多模态资讯展示工具"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "streamlit>=1.35",
    "openai>=1.30",
    "httpx>=0.27",
    "trafilatura>=1.8",
    "python-pptx>=0.6.23",
    "pillow>=10.3",
    "pydantic-settings>=2.2",
    "beautifulsoup4>=4.12",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
